Moved main code out of main.
